Phase C-1 cascade endpoint: move ResumeGuardDescr to majit-backend
ResumeGuardDescr struct + impl Descr + impl FailDescr + helpers
(alloc_fail_index / NEXT_FAIL_INDEX / push_vector_info /
flatten_vector_info / build_vector_info_chain) + status-bit constants
(STATUS_BUSY_FLAG / STATUS_TYPE_MASK / STATUS_SHIFT / STATUS_SHIFT_MASK /
STATUS_TY_NONE/INT/REF/FLOAT) + make_resume_guard_descr_typed factory
move to majit-backend::resume_guard_descr.

All field types (ResumeData, RdPayload, AccumInfo, CompiledLoopToken)
and trait dependencies (Descr, FailDescr) live in majit-ir or
majit-backend after the prior cascade hops.  This is the
unified-descr endpoint: backend codegen can now instantiate
ResumeGuardDescr directly without depending on majit-metainterp.

majit-metainterp::compile re-exports ResumeGuardDescr for caller
compatibility.  Dead #[cfg(any())] placeholder blocks (253 lines)
removed.

This unblocks backend struct deletion: with ResumeGuardDescr
reachable, DynasmFailDescr and CraneliftFailDescr wrapper structs
can be retired in the next cascade hop.
