+++ majit/majit-metainterp/src/compile.rs
@@ -2904,260 +2904,13 @@ pub(crate) const STATUS_TY_FLOAT: u64 = 0x06;
 // where placeholder guard descrs are the same `ResumeGuardDescr`
 // instances backfilled by `store_final_boxes_in_guard` (compile.py:869).
 
-/// Per-guard FailDescr that also carries resume data for deoptimization.
-///
-/// Mirrors RPython's ResumeGuardDescr with snapshot information.
-/// When a guard fails, the backend uses the resume data to reconstruct
-/// the interpreter state (virtual objects, frame variables, etc.).
-#[derive(Debug)]
-struct ResumeGuardDescr {
-    fail_index: u32,
-    /// `compile.py:869 store_final_boxes` mutates types in place; pyre
-    /// uses `UnsafeCell` so identity is preserved across the optimizer.
-    types: UnsafeCell<Vec<Type>>,
-    /// Pyre keeps `resume_data` (the RPython-style ResumeValueSource
-    /// payload used by `prepare_pendingfields` for the
-    /// `PendingFieldInfo` path) on the descr alongside the RPython
-    /// `_attrs_` rd_* slots — both representations co-exist while
-    /// the runtime resume reader is being aligned with upstream.
-    resume_data: ResumeData,
-    /// `compile.py:855` `_attrs_ = ('rd_numb', 'rd_consts',
-    /// 'rd_virtuals', 'rd_pendingfields', 'status')`.
-    payload: RdPayload,
-    /// RPython history.py:127 rd_vector_info — no Mutex needed, single-threaded.
-    vector_info: UnsafeCell<Option<Box<AccumInfo>>>,
-    /// `compile.py:186` `descr.rd_loop_token = clt` line-by-line port.
-    ///
-    /// The owning `Arc<CompiledLoopToken>` typed as `Arc<dyn Any +
-    /// Send + Sync>` (the `token_handle_any` pattern — `majit-ir`
-    /// cannot reference `majit-backend::CompiledLoopToken` without a
-    /// dependency cycle).  Stamped by `record_loop_or_bridge`
-    /// (`pyjitpl/mod.rs::record_loop_or_bridge`, compile.py:183-186
-    /// counterpart) for every resume guard, alongside (and
-    /// transitionally with) the same stamp on the backend descr.
-    /// `pyjitpl.py:2897 resumedescr.rd_loop_token.loop_token_wref()`
-    /// readers chain through this slot; `clt.loop_token_wref()`
-    /// (compile.py:180-181) yields the owning `JitCellToken` weakref.
-    rd_loop_token_clt: UnsafeCell<Option<std::sync::Arc<CompiledLoopToken>>>,
-    /// `history.py:132` `AbstractFailDescr._attrs_ = ('adr_jump_offset',
-    /// 'rd_locs', 'rd_loop_token', 'rd_vector_info')`.
-    ///
-    /// `assembler.py:849 patch_pending_failure_recoveries` stamps the
-    /// absolute address of the recovery stub.  `assembler.py:987
-    /// patch_jump_for_descr` resets to `0` after a bridge is attached
-    /// ("0 means patched").  `UnsafeCell` so identity is preserved
-    /// across the write — the `Arc<dyn FailDescr>` address stamped on
-    /// the IR op stays valid.
-    adr_jump_offset: UnsafeCell<usize>,
-    /// `history.py:132` `AbstractFailDescr._attrs_` `rd_locs`.  Written
-    /// by `llsupport/assembler.py:279 guardtok.faildescr.rd_locs =
-    /// positions`; read by `llsupport/llmodel.py:424
-    /// descr.rd_locs[index] * WORD`.  Empty until codegen stamps it.
-    rd_locs: UnsafeCell<Vec<u16>>,
-    /// `compile.py:683` `AbstractResumeGuardDescr._attrs_ = ('status',)`.
-    /// Packs `ST_BUSY_FLAG` + type tag + jitcounter hash;
-    /// `start_compiling` / `done_compiling` toggle the busy bit and
-    /// `store_hash` / `make_a_counter_per_value` write the rest.
-    status: AtomicU64,
-    /// Pyre-only: identifier of the compiled trace that owns this guard.
-    ///
-    /// RPython resolves descr identity directly by Python `id(descr)`
-    /// (`history.py:125`).  Pyre's runtime exit path uses a
-    /// `(trace_id, fail_index)` key when threading guard failures back
-    /// from `runner.rs::find_descr` / `compiler.rs::find_descr_by_ptr`,
-    /// so the owning trace_id is captured on the descr by
-    /// `record_loop_or_bridge` (`compile.py:185-186` walker) and read
-    /// directly via `FailDescr::trace_id`.
-    trace_id: AtomicU64,
-    /// Pyre-only: per-trace `fail_index` assigned by
-    /// `build_guard_metadata`.  The struct's `fail_index` field above
-    /// holds the global `alloc_fail_index()` value used for descr
-    /// identity; this slot mirrors the per-trace key the optimizer
-    /// hands the backend (`(trace_id, fail_index)` lookup parity).
-    /// Stamped at `build_guard_metadata` time, where the local
-    /// `fail_index` counter increments per guard.
-    fail_index_per_trace: AtomicU32,
-}
+// ResumeGuardDescr moved to majit-backend::resume_guard_descr (Phase
+// C-1 cascade endpoint).  Re-exported here for caller compatibility;
+// all impl Descr / impl FailDescr live in majit-backend along with
+// the helpers (alloc_fail_index, push_vector_info, etc.) and status
+// constants.
+pub use majit_backend::ResumeGuardDescr;
 
-unsafe impl Send for ResumeGuardDescr {}
-unsafe impl Sync for ResumeGuardDescr {}
-
-impl majit_ir::Descr for ResumeGuardDescr {
-    fn index(&self) -> u32 {
-        self.fail_index
-    }
-    fn as_fail_descr(&self) -> Option<&dyn FailDescr> {
-        Some(self)
-    }
-    fn is_resume_guard(&self) -> bool {
-        true
-    }
-    /// compile.py:844-846: ResumeGuardDescr.clone()
-    fn clone_descr(&self) -> Option<DescrRef> {
-        Some(Arc::new(ResumeGuardDescr {
-            fail_index: alloc_fail_index(),
-            types: UnsafeCell::new(unsafe { (&*self.types.get()).clone() }),
-            resume_data: self.resume_data.clone(),
-            payload: self.payload.deep_clone(),
-            vector_info: UnsafeCell::new(unsafe { (&*self.vector_info.get()).clone() }),
-            // `compile.py:844-846` `ResumeGuardDescr()` mints a default-
-            // attributes object; the `_attrs_` slots (adr_jump_offset /
-            // rd_locs / status / rd_loop_token / rd_vector_info) reset
-            // to their initial values when this fresh descr reaches
-            // backend codegen (`llsupport/assembler.py:279`).
-            adr_jump_offset: UnsafeCell::new(0),
-            rd_locs: UnsafeCell::new(Vec::new()),
-            status: AtomicU64::new(0),
-            rd_loop_token_clt: UnsafeCell::new(None),
-            // Cloned descrs reach the backend codegen path freshly, so
-            // `record_loop_or_bridge` re-stamps `trace_id` for the
-            // owning compiled trace.
-            trace_id: AtomicU64::new(0),
-            fail_index_per_trace: AtomicU32::new(0),
-        }))
-    }
-}
-
-impl FailDescr for ResumeGuardDescr {
-    fn fail_index(&self) -> u32 {
-        // `assembler.py:227 self.faildescr.index = i` — `descr.fail_index`
-        // is the PER-TRACE key the backend stamps at codegen, NOT the
-        // global identity counter (the global `alloc_fail_index()`
-        // value in `self.fail_index` is pyre-only serialization
-        // identity, accessible via `Descr::index()`).
-        self.fail_index_per_trace.load(Ordering::Relaxed)
-    }
-    fn trace_id(&self) -> u64 {
-        self.trace_id.load(Ordering::Relaxed)
-    }
-    fn set_trace_id(&self, trace_id: u64) {
-        self.trace_id.store(trace_id, Ordering::Relaxed);
-    }
-    fn fail_index_per_trace(&self) -> u32 {
-        self.fail_index_per_trace.load(Ordering::Relaxed)
-    }
-    fn set_fail_index_per_trace(&self, fail_index: u32) {
-        self.fail_index_per_trace
-            .store(fail_index, Ordering::Relaxed);
-    }
-    fn fail_arg_types(&self) -> &[Type] {
-        unsafe { &*self.types.get() }
-    }
-    fn set_fail_arg_types(&self, types: Vec<Type>) {
-        unsafe { *self.types.get() = types }
-    }
-    fn attach_vector_info(&self, info: AccumInfo) {
-        push_vector_info(unsafe { &mut *self.vector_info.get() }, info);
-    }
-    fn vector_info(&self) -> Vec<AccumInfo> {
-        flatten_vector_info(unsafe { (&*self.vector_info.get()).as_deref() })
-    }
-    fn replace_vector_info(&self, chain: Vec<AccumInfo>) {
-        unsafe { *self.vector_info.get() = build_vector_info_chain(chain) }
-    }
-
-    fn rd_numb(&self) -> Option<&[u8]> {
-        self.payload.rd_numb()
-    }
-    fn rd_numb_arc(&self) -> Option<Arc<[u8]>> {
-        self.payload.rd_numb_arc()
-    }
-    fn set_rd_numb(&self, value: Option<Vec<u8>>) {
-        self.payload.set_rd_numb(value)
-    }
-    fn set_rd_numb_arc(&self, value: Option<Arc<[u8]>>) {
-        self.payload.set_rd_numb_arc(value)
-    }
-    fn rd_consts(&self) -> Option<&[Const]> {
-        self.payload.rd_consts()
-    }
-    fn rd_consts_arc(&self) -> Option<Arc<[Const]>> {
-        self.payload.rd_consts_arc()
-    }
-    fn set_rd_consts(&self, value: Option<Vec<Const>>) {
-        self.payload.set_rd_consts(value)
-    }
-    fn set_rd_consts_arc(&self, value: Option<Arc<[Const]>>) {
-        self.payload.set_rd_consts_arc(value)
-    }
-    fn rd_virtuals(&self) -> Option<&[Rc<RdVirtualInfo>]> {
-        self.payload.rd_virtuals()
-    }
-    fn rd_virtuals_arc(&self) -> Option<Arc<[Rc<RdVirtualInfo>]>> {
-        self.payload.rd_virtuals_arc()
-    }
-    fn set_rd_virtuals(&self, value: Option<Vec<Rc<RdVirtualInfo>>>) {
-        self.payload.set_rd_virtuals(value)
-    }
-    fn set_rd_virtuals_arc(&self, value: Option<Arc<[Rc<RdVirtualInfo>]>>) {
-        self.payload.set_rd_virtuals_arc(value)
-    }
-    fn rd_pendingfields(&self) -> Option<&[GuardPendingFieldEntry]> {
-        self.payload.rd_pendingfields()
-    }
-    fn rd_pendingfields_arc(&self) -> Option<Arc<[GuardPendingFieldEntry]>> {
-        self.payload.rd_pendingfields_arc()
-    }
-    fn set_rd_pendingfields(&self, value: Option<Vec<GuardPendingFieldEntry>>) {
-        self.payload.set_rd_pendingfields(value)
-    }
-    fn set_rd_pendingfields_arc(&self, value: Option<Arc<[GuardPendingFieldEntry]>>) {
-        self.payload.set_rd_pendingfields_arc(value)
-    }
-    fn adr_jump_offset(&self) -> usize {
-        // Safety: single-threaded JIT, no concurrent writers.
-        unsafe { *self.adr_jump_offset.get() }
-    }
-    fn set_adr_jump_offset(&self, offset: usize) {
-        // Safety: single-threaded JIT, no concurrent readers.
-        unsafe { *self.adr_jump_offset.get() = offset };
-    }
-    fn rd_locs(&self) -> &[u16] {
-        // Safety: single-threaded JIT, no concurrent writers.
-        unsafe { &*self.rd_locs.get() }
-    }
-    fn set_rd_locs(&self, locs: Vec<u16>) {
-        // Safety: single-threaded JIT, no concurrent readers.
-        unsafe { *self.rd_locs.get() = locs };
-    }
-    fn get_status(&self) -> u64 {
-        self.status.load(Ordering::Acquire)
-    }
-    fn start_compiling(&self) {
-        self.status
-            .fetch_or(STATUS_BUSY_FLAG, Ordering::AcqRel);
-    }
-    fn done_compiling(&self) {
-        self.status
-            .fetch_and(!STATUS_BUSY_FLAG, Ordering::AcqRel);
-    }
-    fn store_hash(&self, hash: u64) {
-        self.status
-            .store(hash & STATUS_SHIFT_MASK, Ordering::Release);
-    }
-    fn make_a_counter_per_value(&self, index: u32, type_tag: u64) {
-        let value = type_tag | ((index as u64) << STATUS_SHIFT);
-        self.status.store(value, Ordering::Release);
-    }
-    fn rd_loop_token_clt(&self) -> Option<&dyn std::any::Any> {
-        // compile.py:186 reader: return `&Arc<CompiledLoopToken>` typed
-        // through `dyn Any` so `descr_owning_clt` can `downcast_ref::<Arc<
-        // CompiledLoopToken>>()` consistently across backend + metainterp
-        // descrs.  Safety: single-threaded JIT, no concurrent writers.
-        let cell = unsafe { &*self.rd_loop_token_clt.get() };
-        cell.as_ref().map(|arc| arc as &dyn std::any::Any)
-    }
-    fn set_rd_loop_token_clt(&self, clt: std::sync::Arc<dyn std::any::Any + Send + Sync>) {
-        // compile.py:186 stamp.  Downcast at storage time so the cell
-        // holds the strongly typed `Arc<CompiledLoopToken>` matching
-        // backend descrs.  Safety: single-threaded JIT, no concurrent
-        // readers.
-        let typed: std::sync::Arc<CompiledLoopToken> = clt
-            .downcast::<CompiledLoopToken>()
-            .expect("set_rd_loop_token_clt expected Arc<CompiledLoopToken>");
-        unsafe { *self.rd_loop_token_clt.get() = Some(typed) };
-    }
-}
 
 /// Create a FailDescr for `num_live` integer values with an auto-assigned
 /// unique fail_index.
